backend/requirements.txt
Flask
Flask-CORS
torch
torchvision
Pillow
numpy
psycopg2-binary
efficientnet-pytorch
python-dotenv